singularized file names, moved primitive q/m into scene
