+++ src/gql/queries/asset.ts
