+++ src/gql/queries/index.ts
@@ -1,14 +1,13 @@
-export * from "./assets";
-export * from "./clusters";
-export * from "./datasets";
+export * from "./asset";
+export * from "./cluster";
+export * from "./dataset";
 export * from "./infobox";
-export * from "./layers";
-export * from "./plugins";
-export * from "./primitives";
-export * from "./projects";
-export * from "./properties";
+export * from "./layer";
+export * from "./plugin";
+export * from "./project";
+export * from "./property";
 export * from "./scene";
-export * from "./tags";
+export * from "./tag";
 export * from "./user";
-export * from "./widgets";
+export * from "./widget";
 export * from "./workspace";
