+++ src/gql/queries/plugin.ts
