+++ src/gql/queries/property.ts
