+++ src/gql/queries/widget.ts
